requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CrewAI
crewai==0.30.0
crewai-tools==0.4.0

# LLM
openai==1.12.0
langchain==0.1.0
langchain-openai==0.0.5

# Database
supabase==2.3.0

# Web Scraping
firecrawl-py==0.0.16

# HTTP Client
httpx==0.26.0

# Google APIs
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# PDF Generation
weasyprint==60.2
markdown==3.5.2
jinja2==3.1.2

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
